refactor: rename style object
