+++ src/components/SignUp/Terms/TermsForm/termsForm.css.ts
@@ -2,7 +2,7 @@ import { vars } from "@/styles/theme.css";
 import { style } from "@vanilla-extract/css";
 import { recipe } from "@vanilla-extract/recipes";
 
-export const agreeContainer = style({
+export const termsFormWrapper = style({
   flex: 1,
 
   display: "flex",
@@ -11,39 +11,11 @@ export const agreeContainer = style({
   gap: "2rem",
 });
 
-export const optionalAgreeContainer = style({
+export const termCheckboxListContainer = style({
   display: "flex",
   flexDirection: "column",
   gap: "1.8rem",
   padding: "0 2rem",
 
   marginTop: "2.7rem",
-});
-
-export const optionalAgreeWrapper = style({
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "space-between",
-});
-
-export const agreeWrapper = style({
-  display: "flex",
-  alignItems: "center",
-  gap: "1.5rem",
-});
-
-export const agreeText = recipe({
-  base: {
-    ...vars.fontStyles.body2_m_14,
-  },
-  variants: {
-    isRequired: {
-      true: {
-        color: vars.color.mint_500,
-      },
-      false: {
-        color: vars.color.grey_6,
-      },
-    },
-  },
 });
